+++ LiteLoader.NET/Header/RemoteCall/include.hpp
@@ -15,8 +15,6 @@
 #include "../../Header/MC/Vec3.hpp"
 #include "../../Header/MC/BlockPos.hpp"
 
-#include "../../Header/Logger/Logger.hpp"
-
 //#define REMOTECALL_DEBUG
 
 #define REMOTECALL_DEBUG_INFO(info)							\
