+++ LiteLoader.NET/Kernel/FormUI.cpp
@@ -2,8 +2,6 @@
 
 #include "../Header/MC/Player.hpp"
 
-#include "../Header/Logger/Logger.hpp"
-
 
 namespace LLNET::Form
 {
