+++ LiteLoader.NET/Kernel/MC/BoundingBox.cpp
@@ -4,8 +4,6 @@
 
 #include "../../Tools/NativeCallbackConverter.hpp"
 
-#include <LiteLoader.NET/Header/Logger/Logger.hpp>
-
 
 namespace MC {
 	DelegateToNativeHelper(ForEachBlockInBoxHelper, BoundingBox::ForEachBlockInBoxHandler, void, ::BlockPos const& pos) {
