+++ LiteLoader.NET/Kernel/MC/Player.cpp
@@ -13,11 +13,9 @@
 #include "../../Header/MC/AABB.hpp"
 #include "../../Header/MC/Tick.hpp"
 
+#include "../../Header/Logger.hpp"
 #include "LiteLoader.NET/Header/MC/Packet.hpp"
 #include "LiteLoader.NET/Header/MC/SerializedSkin.hpp"
-
-#include <LiteLoader.NET/Header/Logger/Logger.hpp>
-
 
 enum GameType;
