+++ LiteLoader.NET/LiteLoader.NET.vcxproj.filters
@@ -76,12 +76,6 @@
     <Filter Include="Extra\net_native">
       <UniqueIdentifier>{715adb0f-1875-4f7b-9dbd-f3e0d50d1d33}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Header\Logger">
-      <UniqueIdentifier>{0e00fe85-d6d7-48d6-a04a-ddbde7f40b60}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Kernel\Logger">
-      <UniqueIdentifier>{5eed7d25-5396-4592-9ec4-4b4b45f12f88}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Kernel\MC\ActorDamageSource.cpp">
@@ -196,6 +190,9 @@
       <Filter>Kernel</Filter>
     </ClCompile>
     <ClCompile Include="Kernel\LLAPI.cpp">
+      <Filter>Kernel</Filter>
+    </ClCompile>
+    <ClCompile Include="Kernel\Logger.cpp">
       <Filter>Kernel</Filter>
     </ClCompile>
     <ClCompile Include="Kernel\PlayerInfo.cpp">
@@ -399,15 +396,6 @@
     <ClCompile Include="Main\GlobalClass.cpp">
       <Filter>Main</Filter>
     </ClCompile>
-    <ClCompile Include="Kernel\Logger\Logger.cpp">
-      <Filter>Kernel\Logger</Filter>
-    </ClCompile>
-    <ClCompile Include="Kernel\Logger\LoggerManager.cpp">
-      <Filter>Kernel\Logger</Filter>
-    </ClCompile>
-    <ClCompile Include="Extra\i18nQAQ.cpp">
-      <Filter>Extra</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Header\MC\ActorDamageSource.hpp">
@@ -528,6 +516,9 @@
       <Filter>Header</Filter>
     </ClInclude>
     <ClInclude Include="Header\LLAPI.h">
+      <Filter>Header</Filter>
+    </ClInclude>
+    <ClInclude Include="Header\Logger.hpp">
       <Filter>Header</Filter>
     </ClInclude>
     <ClInclude Include="Header\PlayerInfo.h">
@@ -865,12 +856,6 @@
     </ClInclude>
     <ClInclude Include="Main\FixCLRFatalError.h">
       <Filter>Main</Filter>
-    </ClInclude>
-    <ClInclude Include="Header\Logger\Logger.hpp">
-      <Filter>Header\Logger</Filter>
-    </ClInclude>
-    <ClInclude Include="Header\Logger\LoggerManager.hpp">
-      <Filter>Header\Logger</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
