+++ LiteLoader.NET/Main/GlobalClass.cpp
@@ -1,6 +1,5 @@
 #include "GlobalClass.h"
 
-#include <LiteLoader.NET/Header/Logger/Logger.hpp>
 
 inline void GlobalClass::Init()
 {
