+++ LiteLoader.NET/Main/Version.h
@@ -7,6 +7,7 @@
 
 
 
+
 #define LLNET_ASSEMBLY_VERSION  TO_VERSION_STRING(LLNET_VERSION_MAGOR.LLNET_VERSION_MINOR.LLNET_VERSION_REVISION.*)
 
 #define LLNET_ASSEMBLY_FILE_VERSION  TO_VERSION_STRING(LLNET_VERSION_MAGOR.LLNET_VERSION_MINOR.LLNET_VERSION_REVISION.LLNET_VERSION_BUILD)
